requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.59.3
pillow==11.0.0
aiofiles==24.1.0
